% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\docType{data}
\name{age}
\alias{age}
\title{Ages of all Williams faculty}
\format{A data frame with 403 rows and 4 variables:
\describe{
   \item{name}{the professor's name}
   \item{department}{the professor's department}
   \item{year}{the year the professor graudated college}
   \item{age}{the age of the professor, in years}
 }
 @source \url{http://web.williams.edu/admin/registrar//catalog/archive.html}}
\usage{
age
}
\description{
A dataset containing the department, graduation year, and age of all 403 Williams
faculty, including visiting lecturers
}
\keyword{datasets}

