% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_summary.R
\name{age_summary}
\alias{age_summary}
\title{Generates a summary of relevant data}
\usage{
age_summary()
}
\value{
data frame containing summary of relevant data
}
\description{
This functions creates and prints a data frame containing relevant information
about the data, including the total number of professors, the age difference between
the oldest and youngest professor, the name of the oldest professor, the name of the
youngest professor, the average age, the standard deviation of the ages, and the
variance of the ages.
}
\examples{
age_summary()
}

