% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_age.R
\name{average_age}
\alias{average_age}
\title{Calculates the average age of all Williams faculty}
\usage{
average_age()
}
\value{
the average age of all Williams faculty
}
\description{
This function simply calculates and returns the mean of the ages of all Williams faculty.
}
\examples{
average_age()

}

