% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_by_dpmt.R
\name{color_by_dpmt}
\alias{color_by_dpmt}
\title{Visualizing professor ages by department}
\usage{
color_by_dpmt(dpmt_displayed)
}
\arguments{
\item{dpmt_displayed}{This input should include the names of the departments desired, in the form of a list}
}
\value{
a colored scatter plot of Williams faculty ages by department
}
\description{
This function creates a colored scatter plot with each color representing a different
department. The index represents the order of the professors in the original data, since
the original data is in alphabetical order, the index gives a sense of which letter the
professor's last name starts with. A legend is included indicating which color represents which
department. If there is no input, the scatter plot will have points from all 22 departments, with the oldest
and youngest professors pointed out. If an input is given, the scatter plot will only contain points for the professors in the specified
departments. If the input contains a string that is not a department, an error message will be printed.
Note: the legend may overlap with the plot in some RStudio layouts, adjust the size of the plot area to 
default for clear graph.
}
\examples{
color_by_dpmt()
color_by_dpmt(c('Theatre/Dance','Biology/Environmental Studies', 'Anthropology/Sociology'))
}
\seealso{
\code{\link{plot_age}} for histogram of all ages, \code{\link{plot_by_dpmt}}
for histograms by department, \code{\link{plot_by_gender}} for histograms by gender, and
\code{\link{color_by_gender}} for colored scatter plot by gender

Other plotting.functions: \code{\link{color_by_gender}},
  \code{\link{plot_age}}, \code{\link{plot_by_dpmt}},
  \code{\link{plot_by_gender}}
}

