% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_by_gender.R
\name{color_by_gender}
\alias{color_by_gender}
\title{Visualizing professor ages by gender}
\usage{
color_by_gender()
}
\value{
a colored scatter plot of Williams faculty ages by gender
}
\description{
This function creates a colored scatter plot with each color representing a different
gender. The index represents the order of the professors in the original data, since
the original data is in alphabetical order, the index gives a sense of which letter the
professor's last name starts with. A legend is included indicating which color represents which
gender.
}
\examples{
color_by_gender()
}
\seealso{
\code{\link{plot_age}} for histogram of all ages, \code{\link{plot_by_dpmt}}
for histograms by department, \code{\link{plot_by_gender}} for histograms by gender, and
\code{\link{color_by_dpmt}} for colored scatter plot by department

Other plotting.functions: \code{\link{color_by_dpmt}},
  \code{\link{plot_age}}, \code{\link{plot_by_dpmt}},
  \code{\link{plot_by_gender}}
}

