% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_table.R
\name{dt_table}
\alias{dt_table}
\title{Generates a DT table of the data}
\usage{
dt_table()
}
\value{
a DT data table of the faculty age data
}
\description{
This functions simply creates a DT data table for all the data. The DT table
is useful for the user to explore and manipulate the data.
}

