% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_age.R
\name{plot_age}
\alias{plot_age}
\title{Plots the age distribution of all Williams faculty}
\usage{
plot_age()
}
\value{
a histogram of the age distribution of all Williams faculty

a histogram of the age distribution of all Williams faculty
}
\description{
This functions creates a histogram of the age distribution of all Williams faculty.
}
\examples{
plot_age()
}
\seealso{
\code{\link{plot_by_dpmt}} for histograms by department, \code{\link{plot_by_gender}}
for histograms by gender,

Other plotting.functions: \code{\link{color_by_dpmt}},
  \code{\link{color_by_gender}},
  \code{\link{plot_by_dpmt}}, \code{\link{plot_by_gender}}
}

