% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_age_distribution}
\alias{plot_age_distribution}
\title{Plots the age distribution of all Williams faculty}
\usage{
plot_age_distribution()
}
\description{
Plots the age distribution of all Williams faculty
}

