% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_by_dpmt.R
\name{plot_by_dpmt}
\alias{plot_by_dpmt}
\title{Plotting age by department}
\usage{
plot_by_dpmt(dpmt)
}
\arguments{
\item{dpmt}{This input should be one of the departments at Williams, the first letter
should be capitalized}
}
\value{
a histogram of the age distribution of Williams faculty of the department
selected
}
\description{
This function generates plots of takes an input, and if that input is a department, it will create
a histogram of the age distribution of Williams faculty of that department. If
the input is not a department, an error message would be returned. If the input
is one of the department in a combined category, a histogram of that category
would be returned.
}
\examples{
plot_by_dpmt('Physics')
plot_by_dpmt('Theatre/Dance')
}
\seealso{
\code{\link{plot_age}} for histogram of all ages, \code{\link{plot_by_gender}}
for histograms by gender, \code{\link{color_by_gender}} for colored scatter plot by gender, and
\code{\link{color_by_dpmt}} for colored scatter plot by department

Other plotting.functions: \code{\link{color_by_dpmt}},
  \code{\link{color_by_gender}}, \code{\link{plot_age}},
  \code{\link{plot_by_gender}}
}

