% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_by_gender.R
\name{plot_by_gender}
\alias{plot_by_gender}
\title{Plotting age by gender}
\usage{
plot_by_gender(gender)
}
\arguments{
\item{gender}{This input should be one of the genders}
}
\value{
a histogram of the age distribution of Williams faculty of the gender selected
}
\description{
This function takes an input, and if that input is a gender, it will create
a histogram of the age distribution of Williams faculty of that gender. If
the input is not a gender, an error message would be returned.
}
\examples{
plot_by_gender("Male")
plot_by_gender("female")
}
\seealso{
\code{\link{plot_age}} for histogram of all ages, \code{\link{plot_by_dpmt}}
for histograms by department,

Other plotting.functions: \code{\link{plot_age}},
  \code{\link{plot_by_dpmt}}
}

