% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_age.R
\name{test_age}
\alias{test_age}
\title{Statistical analysis on data}
\usage{
test_age(dpmt = TRUE, gender = TRUE)
}
\description{
This function runs statistical analysis on the data. The tests run include a binomial
test for the number of female professors, a two-sample t-test with male and female professor
ages, and a one-way ANOVA with professor ages in each department.
}

