% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/williamsFaculty-package.R
\docType{package}
\name{williamsFaculty-package}
\alias{williamsFaculty-package}
\title{A package to visualize data on Williams faculty age}
\description{
\tabular{ll}{ Package \tab williamsFaculty \cr Type: \tab Package\cr 
Version: \tab 1.0\cr Date: \tab 2016-02-03\cr License: \tab GPL-3\cr }
}
\details{
The williamsFaculty package focuses on the ages of Williams College faculty members.
The package
}

