% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_summary.R
\name{age_summary}
\alias{age_summary}
\title{Generates a summary of relevant data}
\usage{
age_summary(dpmt = FALSE, gender = FALSE)
}
\arguments{
\item{dpmt}{This argument is initialized as FALSE. When set to TRUE, the function generates
summary of data with regards to departments}

\item{gender}{This argument is initialized as FALSE. Wehn set to TRUE, the function generates 
summary of data with regards to gender}
}
\value{
summaries of data depending on input
}
\description{
This functions generates summaries of the data. If neither 'dpmt' or 'gender'
is specified as TRUE, the function returns a data frame with attributes of the whole dataframe.
If 'gender' is set to true, the function returns a data frame with attributes of the data with
regards to gender. If 'dpmt' is set to true, the function generates a DT data table with attributes of
the data with regards to departments.
}
\examples{
age_summary()
age_summary(dpmt=TRUE)
age_summary(gender=TRUE)
}

