% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WilliamsFaculty.R
\name{average_age}
\alias{average_age}
\title{Calculates the mean of the ages of all Williams faculty}
\usage{
average_age()
}
\value{
The average age of all Williams faculty
}
\description{
Calculates the mean of the ages of all Williams faculty
}

