% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_by_gender.R
\name{plot_by_gender}
\alias{plot_by_gender}
\title{Plotting age by gender}
\usage{
plot_by_gender(gender)
}
\arguments{
\item{gender}{This input should be one of the genders}
}
\value{
a histogram of the age distribution of Williams faculty of the gender selected
}
\description{
This function generates plots with gender as a factor. If there is no input, a qplot (boxplots)
of the age data with both genders will be generated. If the input is a gender, it will create
a histogram of the age distribution of Williams faculty of that gender. If
the input is not a gender, an error message would be returned.
}
\examples{
plot_by_gender('Male')
plot_by_gender('female')
}
\seealso{
\code{\link{plot_age}} for histogram of all ages, \code{\link{plot_by_dpmt}}
for boxplots and histograms by department, \code{\link{color_by_gender}} for colored scatter plot by gender, and
\code{\link{color_by_dpmt}} for colored scatter plot by department

Other plotting.functions: \code{\link{color_by_dpmt}},
  \code{\link{color_by_gender}}, \code{\link{plot_age}},
  \code{\link{plot_by_dpmt}}
}

